requirements.txt
dacvae @ git+https://github.com/facebookresearch/dacvae
datasets>=3.0.0
huggingface-hub>=0.34.0,<1.0
llvmlite>=0.40.0
numba>=0.57.0
peft>=0.18.0
pyyaml>=6.0
safetensors>=0.7.0
soundfile>=0.12.0
sentencepiece>=0.1.99,<0.2
torchcodec>=0.10.0
transformers<5
tqdm>=4.67.3
wandb>=0.17.0
